feat(chat): spacer added to push new user message to top & after AI res
The whole layout is replaced with a container with inner scroll to make the general html/body tags fixed and non-scrollable.
When new messages are added to the container, the user message has to jump to the perfect top.
If the user message is too tall, only last 3 lines have to be visible at the top.
Also, the spacer is adjusted after AI response because the actual container height is changed and it causes a jump. We need to keep the user message at the same place as it used to be.
All the calculations provided with the @description tag.
Also, usage of Vue Use was required for getting refs of parent container with child messages because the built-in useTemplateRef doesn't guarantee a correct order.
The height of new added DOM elements is not calculated correctly immediately. That's a reason why additional tricks added to handle correct heights with small delay because it's strictly necessary for calculations and correct positioning.